requirements.txt
crewai
crewai-tools
pinecone
openai
streamlit
python-dotenv
requests
beautifulsoup4
lxml
